requirements.txt
matplotlib==3.4.2
networkx==2.5.1
dimod==0.10.3
dwave-system==1.7.0
jupyterlab==3.0.16